backend/pyproject.toml
[tool.poetry]
name = "Tasky"
version = "0.1.0"
description = ""
authors = ["MelDxKviel <mega1hack52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.4.0"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
